Migrate to PostgreSQL

+++ model/sql/reminders.go
@@ -3,8 +3,9 @@ package sql
 import (
 	"database/sql"
 	"errors"
-	"github.com/PaulSonOfLars/gotgbot/v2"
 	"time"
+
+	"github.com/PaulSonOfLars/gotgbot/v2"
 
 	"github.com/Brawl345/gobot/logger"
 	"github.com/Brawl345/gobot/model"
@@ -27,10 +28,10 @@ func (db *reminderService) DeleteReminder(chat *gotgbot.Chat, user *gotgbot.User
 	var exists bool
 	var err error
 	if chat.Type == gotgbot.ChatTypePrivate {
-		const existsQuery = `SELECT 1 FROM reminders WHERE id = ? AND chat_id IS NULL AND user_id = ?`
+		const existsQuery = `SELECT 1 FROM reminders WHERE id = $1 AND chat_id IS NULL AND user_id = $2`
 		err = db.Get(&exists, existsQuery, id, user.Id)
 	} else {
-		const existsQuery = `SELECT 1 FROM reminders WHERE id = ? AND chat_id = ?`
+		const existsQuery = `SELECT 1 FROM reminders WHERE id = $1 AND chat_id = $2`
 		err = db.Get(&exists, existsQuery, id, chat.Id)
 	}
 
@@ -46,10 +47,10 @@ func (db *reminderService) DeleteReminder(chat *gotgbot.Chat, user *gotgbot.User
 	}
 
 	if chat.Type == gotgbot.ChatTypePrivate {
-		const query = `DELETE FROM reminders WHERE id = ? AND chat_id IS NULL AND user_id = ?`
+		const query = `DELETE FROM reminders WHERE id = $1 AND chat_id IS NULL AND user_id = $2`
 		_, err = db.Exec(query, id, user.Id)
 	} else {
-		const query = `DELETE FROM reminders WHERE id = ? AND chat_id = ?`
+		const query = `DELETE FROM reminders WHERE id = $1 AND chat_id = $2`
 		_, err = db.Exec(query, id, chat.Id)
 	}
 
@@ -57,7 +58,7 @@ func (db *reminderService) DeleteReminder(chat *gotgbot.Chat, user *gotgbot.User
 }
 
 func (db *reminderService) DeleteReminderByID(id int64) error {
-	const query = `DELETE FROM reminders WHERE id = ?`
+	const query = `DELETE FROM reminders WHERE id = $1`
 	_, err := db.Exec(query, id)
 	return err
 }
@@ -72,7 +73,7 @@ func (db *reminderService) GetAllReminders() ([]model.Reminder, error) {
 func (db *reminderService) GetReminderByID(id int64) (model.Reminder, error) {
 	const query = `SELECT chat_id, user_id, username, text FROM reminders r 
     RIGHT JOIN users u ON r.user_id = u.id
-	WHERE r.id = ?`
+	WHERE r.id = $1`
 	var reminder model.Reminder
 	err := db.Get(&reminder, query, id)
 	if err != nil {
@@ -89,10 +90,10 @@ func (db *reminderService) GetReminders(chat *gotgbot.Chat, user *gotgbot.User) 
 	var reminders []model.Reminder
 
 	if chat.Type == gotgbot.ChatTypePrivate {
-		const query = `SELECT id, time, text FROM reminders WHERE chat_id IS NULL AND user_id = ? ORDER BY time`
+		const query = `SELECT id, time, text FROM reminders WHERE chat_id IS NULL AND user_id = $1 ORDER BY time`
 		err = db.Select(&reminders, query, user.Id)
 	} else {
-		const query = `SELECT id, time, text FROM reminders WHERE chat_id = ? ORDER BY time`
+		const query = `SELECT id, time, text FROM reminders WHERE chat_id = $1 ORDER BY time`
 		err = db.Select(&reminders, query, chat.Id)
 	}
 
@@ -106,17 +107,16 @@ func (db *reminderService) SaveReminder(
 	text string,
 ) (int64, error) {
 	var err error
-	var res sql.Result
+	var lastInsertID int64
 	if chat.Type == gotgbot.ChatTypePrivate {
-		const query = `INSERT INTO reminders (user_id, time, text) VALUES (?, ?, ?)`
-		res, err = db.Exec(query, user.Id, remindAt, text)
+		const query = `INSERT INTO reminders (user_id, time, text) VALUES ($1, $2, $3) RETURNING id`
+		err = db.QueryRow(query, user.Id, remindAt, text).Scan(&lastInsertID)
 	} else {
-		const query = `INSERT INTO reminders (chat_id, user_id, time, text) VALUES (?, ?, ?, ?)`
-		res, err = db.Exec(query, chat.Id, user.Id, remindAt, text)
+		const query = `INSERT INTO reminders (chat_id, user_id, time, text) VALUES ($1, $2, $3, $4) RETURNING id`
+		err = db.QueryRow(query, chat.Id, user.Id, remindAt, text).Scan(&lastInsertID)
 	}
 	if err != nil {
 		return 0, err
 	}
-	lastInsertedID, err := res.LastInsertId()
-	return lastInsertedID, err
+	return lastInsertID, err
 }
